chore: move _Types in pddl._validation
